feat: image processing GoL
